requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.2.0
celery==4.4.0
Django==3.0.3
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.0.1
kombu==4.6.7
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
vine==1.3.0
